+++ src/libft.h
@@ -6,13 +6,13 @@
 /*   By: lphelipe <lphelipe@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/04 23:37:42 by lphelipe          #+#    #+#             */
-/*   Updated: 2022/09/04 23:39:42 by lphelipe         ###   ########.fr       */
+/*   Updated: 2022/09/05 01:12:06 by lphelipe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 
-void	*ft_memset(void *b, int c, int len);
+void	*ft_memset(void *b, int c, size_t len);
 int		ft_isalpha(int c);
 #endif
